init master branch

+++ http.h
@@ -3,4 +3,4 @@
 
 struct http_file_info *sound_file_open(const char *URL);
 int http_file_close(struct sound_file_into *file);
-size_t http_file_load(struct sound_file_into *file, u8 *buff, size_t size);
+int http_file_load(struct sound_file_into *file, u8 *buff, size_t size);
